refactor: mv LoraReceiver files to common dir
