+++ applications_user/lora/lora_state_manager.c
@@ -1,5 +1,5 @@
 #include "lora_state_manager.h"
-#include "views/lora_receiver.h"
+#include "lora_receiver.h"
