+++ applications_user/lora/tests/unit_tests/parsers_test.c
@@ -1,7 +1,7 @@
 #include <furi.h>
 #include <stdbool.h>
 
-#include "../../views/lora_receiver_i.h"
+#include "../../lora_receiver_i.h"
 
 /**
  * Test case for parsing a FPENDING message.
